test(update): chats refactor and add quick profile
