+++ tests/screenobjects/settings/SettingsAboutScreen.ts
