+++ tests/screenobjects/settings/SettingsPrivacyScreen.ts
