+++ tests/specs/reusable-accounts/01-chats-userA.spec.ts
@@ -1,9 +1,14 @@
 import { getUserKey, loginWithTestUser } from "../../helpers/commands";
-import ChatScreen from "../../screenobjects/ChatScreen";
-import FriendsScreen from "../../screenobjects/FriendsScreen";
-import WelcomeScreen from "../../screenobjects/WelcomeScreen";
+import ChatsLayout from "../../screenobjects/chats/ChatsLayout";
+import ComposeAttachment from "../../screenobjects/chats/ComposeAttachment";
+import ContextMenu from "../../screenobjects/chats/ContextMenu";
+import InputBar from "../../screenobjects/chats/InputBar";
+import Messages from "../../screenobjects/chats/Messages";
+import Topbar from "../../screenobjects/chats/Topbar";
+import FriendsScreen from "../../screenobjects/friends/FriendsScreen";
+import WelcomeScreen from "../../screenobjects/welcome-screen/WelcomeScreen";
 
-xdescribe("Two users at the same time - Chat User A", async () => {
+describe("Two users at the same time - Chat User A", async () => {
   it("Load Chat User A account and go to friends screen", async () => {
     // Go to Friends Screen
     await loginWithTestUser();
@@ -33,32 +38,30 @@ xdescribe("Two users at the same time - Chat User A", async () => {
 
   it("Go to chat with friend and wait until user is online", async () => {
     await FriendsScreen.chatWithFriendButton.click();
-    await ChatScreen.waitForIsShown(true);
+    await Topbar.validateTopbarExists();
 
     // Wait until Chat User B is online
-    await ChatScreen.topbarIndicatorOnline.waitForDisplayed({
-      timeout: 240000,
-    });
+    await Topbar.waitUntilRemoteUserIsOnline();
   });
 
   it("Chats - Validate Messages secured text displayed on top of conversation", async () => {
-    await ChatScreen.encryptedMessagesText.waitForDisplayed();
-    await expect(ChatScreen.encryptedMessagesText).toHaveTextContaining(
+    await ChatsLayout.encryptedMessagesText.waitForDisplayed();
+    await expect(ChatsLayout.encryptedMessagesText).toHaveTextContaining(
       "Messages secured by local E2E encryption."
     );
   });
 
   it("Chats - Send a message to the other user", async () => {
-    await ChatScreen.typeMessageOnInput("testing...");
-    await ChatScreen.clickOnSendMessage();
+    await InputBar.typeMessageOnInput("testing...");
+    await InputBar.clickOnSendMessage();
 
-    const textFromMessage = await ChatScreen.getLastMessageSentText();
+    const textFromMessage = await Messages.getLastMessageSentText();
     await expect(textFromMessage).toHaveTextContaining("testing...");
   });
 
   it("Validate Chat Message displays timestamp and user who sent it", async () => {
     //Timestamp from last message sent should be displayed
-    const timeAgo = await ChatScreen.getLastMessageSentTimeAgo();
+    const timeAgo = await Messages.getLastMessageSentTimeAgo();
     await expect(timeAgo).toHaveText(
       /^(?:\d{1,2}\s+(?:second|minute)s?\s+ago|now)$/
     );
@@ -67,66 +70,64 @@ xdescribe("Two users at the same time - Chat User A", async () => {
 
   it("Validate Chat Message sent contents", async () => {
     //Any message you sent yourself should appear within a colored message bubble
-    const lastMessage = await ChatScreen.getLastMessageSentLocator();
+    const lastMessage = await Messages.getLastMessageSentLocator();
     await expect(lastMessage).toBeDisplayed();
   });
 
   it("Validate Chat Message Group displays username picture and online indicator", async () => {
     //Your user image should be displayed next to the message
-    const userImage = await ChatScreen.getLastGroupWrapImage();
+    const userImage = await Messages.getLastGroupWrapImage();
     await expect(userImage).toExist();
 
     //Online indicator of your user should be displayed next to the image
-    const onlineIndicator = await ChatScreen.getLastGroupWrapOnline();
+    const onlineIndicator = await Messages.getLastGroupWrapOnline();
     await expect(onlineIndicator).toExist();
   });
 
   it("Chats - Topbar information", async () => {
     // Validate user image, username and online indicator are displayed on Chat Topbar
-    await expect(ChatScreen.topbarUserImage).toBeDisplayed();
-    await expect(ChatScreen.topbarUserName).toHaveTextContaining("ChatUserB");
-    await expect(ChatScreen.topbarIndicatorOnline).toBeDisplayed();
+    await expect(Topbar.topbarUserImage).toBeDisplayed();
+    await expect(Topbar.topbarUserName).toHaveTextContaining("ChatUserB");
+    await expect(Topbar.topbarIndicatorOnline).toBeDisplayed();
   });
 
   it("Chats - Add user with active chat to Favorites", async () => {
     // Add user to favorites
-    await ChatScreen.addToFavorites();
-    await ChatScreen.favorites.waitForDisplayed();
+    await Topbar.addToFavorites();
+    await Topbar.favorites.waitForDisplayed();
 
     // Favorites Sidebar should be displayed
-    await expect(ChatScreen.favoritesUserImage).toBeDisplayed();
-    await expect(ChatScreen.favoritesUserIndicatorOnline).toBeDisplayed();
-    await expect(ChatScreen.favoritesUserName).toHaveTextContaining(
-      "CHATUSERB"
-    );
+    await expect(Topbar.favoritesUserImage).toBeDisplayed();
+    await expect(Topbar.favoritesUserIndicatorOnline).toBeDisplayed();
+    await expect(Topbar.favoritesUserName).toHaveTextContaining("CHATUSERB");
   });
 
   it("Chats - Remove user with active chat from Favorites", async () => {
     // Remove user from favorites
-    await ChatScreen.removeFromFavorites();
-    await ChatScreen.favorites.waitForExist({ reverse: true });
+    await Topbar.removeFromFavorites();
+    await Topbar.favorites.waitForExist({ reverse: true });
   });
 
   it("Receive Reply - Validate reply message is received from remote user", async () => {
     // Wait until reply is received
-    await ChatScreen.chatMessageReply.waitForDisplayed({ timeout: 180000 });
+    await Messages.chatMessageReply.waitForDisplayed({ timeout: 180000 });
   });
 
   it("Receive Reply - Validate reply message contents", async () => {
     // Validate message replied appears smaller above your reply
-    const replyReceived = await ChatScreen.getLastReplyReceived();
-    const replyReceivedText = await ChatScreen.getLastReplyReceivedText();
+    const replyReceived = await Messages.getLastReplyReceived();
+    const replyReceivedText = await Messages.getLastReplyReceivedText();
     await expect(replyReceived).toBeDisplayed();
     await expect(replyReceivedText).toHaveTextContaining("testing...");
 
     // Validate reply message sent appears as last message
-    const textFromMessage = await ChatScreen.getLastMessageReceivedText();
+    const textFromMessage = await Messages.getLastMessageReceivedText();
     await expect(textFromMessage).toHaveTextContaining("this is a reply");
   });
 
   it("Receive Reply - Validate reply message group contains timestamp", async () => {
     //Timestamp from last message sent should be displayed
-    const timeAgo = await ChatScreen.getLastMessageReceivedTimeAgo();
+    const timeAgo = await Messages.getLastMessageReceivedTimeAgo();
     await expect(timeAgo).toHaveText(
       /^(?:\d{1,2}\s+(?:second|minute)s?\s+ago|now)$/
     );
@@ -135,120 +136,122 @@ xdescribe("Two users at the same time - Chat User A", async () => {
 
   it("Receive Reply - Validate reply message group contains user image and online indicator", async () => {
     //Your user image should be displayed next to the message
-    const userImage = await ChatScreen.getLastGroupWrapImage();
+    const userImage = await Messages.getLastGroupWrapImage();
     await expect(userImage).toExist();
 
     //Online indicator of your user should be displayed next to the image
-    const onlineIndicator = await ChatScreen.getLastGroupWrapOnline();
+    const onlineIndicator = await Messages.getLastGroupWrapOnline();
     await expect(onlineIndicator).toExist();
   });
 
   it("Send two more messages to Chat User B", async () => {
     // Send two messages to Chat User B
-    await ChatScreen.typeMessageOnInput("message two");
-    await ChatScreen.clickOnSendMessage();
+    await InputBar.typeMessageOnInput("message two");
+    await InputBar.clickOnSendMessage();
 
-    await ChatScreen.typeMessageOnInput("message three");
-    await ChatScreen.clickOnSendMessage();
+    await InputBar.typeMessageOnInput("message three");
+    await InputBar.clickOnSendMessage();
   });
 
   it("Context Menu - Delete Message", async () => {
     // Open context menu on last message sent and select option with index = 3 (delete)
-    await ChatScreen.openContextMenuOnSentMessage();
-    await ChatScreen.selectContextOption(3);
+    await Messages.openContextMenuOnSentMessage();
+    await ContextMenu.validateContextMenuIsOpen();
+    await ContextMenu.selectContextOption(3);
 
     // Validate that last message was deleted and therefore the last message displayed is "message two"
-    const textFromMessage = await ChatScreen.getLastMessageSentText();
+    const textFromMessage = await Messages.getLastMessageSentText();
     await expect(textFromMessage).toHaveTextContaining("message two");
   });
 
   it("Context Menu - Edit Message", async () => {
     // Open context menu on last message sent, select option with index = 2 (edit) and type a new message
-    await ChatScreen.openContextMenuOnSentMessage();
-    await ChatScreen.selectContextOption(2);
-    await ChatScreen.typeOnEditMessageInput("edited...");
+    await Messages.openContextMenuOnSentMessage();
+    await ContextMenu.validateContextMenuIsOpen();
+    await ContextMenu.selectContextOption(2);
+    await InputBar.typeOnEditMessageInput("edited...");
 
     // Validate message edited contents is shown on Chat Screen
-    const textFromMessage = await ChatScreen.getLastMessageSentText();
+    const textFromMessage = await Messages.getLastMessageSentText();
     await expect(textFromMessage).toHaveTextContaining("edited...");
   });
 
   it("Message Input - User cannot send empty messages", async () => {
     // Ensure that input bar is empty and click on send message button
-    await ChatScreen.clearInputBar();
-    await ChatScreen.clearInputBar();
-    await ChatScreen.clickOnInputBar();
-    await ChatScreen.clickOnSendMessage();
+    await InputBar.clearInputBar();
+    await InputBar.clearInputBar();
+    await InputBar.clickOnInputBar();
+    await InputBar.clickOnSendMessage();
 
     // Ensure that input bar is empty and press Enter Key
-    await ChatScreen.clearInputBar();
-    await ChatScreen.clickOnInputBar();
-    await ChatScreen.pressEnterKeyOnInputBar();
+    await InputBar.clearInputBar();
+    await InputBar.clickOnInputBar();
+    await InputBar.pressEnterKeyOnInputBar();
 
     // Validate latest message sent displayed on Chat Conversation is still "edited..."
-    const latestMessage = await ChatScreen.getLastMessageSentText();
+    const latestMessage = await Messages.getLastMessageSentText();
     await expect(latestMessage).toHaveTextContaining("edited...");
   });
 
   // Skipping test failing on CI due to slowness on driver typing 1024 characters
   xit("Message Input - User can type up to 1024 chars on input bar", async () => {
     // Generate a random text with 1024 chars
-    const longText = await ChatScreen.generateRandomText();
+    const longText = await InputBar.generateRandomText();
     // Type long text with 1024 chars on input bar and attempt to add 4 more chars (efgh)
-    await ChatScreen.typeMessageOnInput(longText + "efgh");
+    await InputBar.typeMessageOnInput(longText + "efgh");
 
     // Ensure that latest chars were not added to input bar, since the max number of chars has been reached
     // Input bar text should be equal to long text with 1024 chars
-    await expect(ChatScreen.inputText).toHaveText(longText);
+    await expect(InputBar.inputText).toHaveText(longText);
 
     // Clear input bar to finish test
-    await ChatScreen.clearInputBar();
+    await InputBar.clearInputBar();
   });
 
   it("Chats - Validate compose attachments contents", async () => {
     // Click on upload button and attach a file to compose attachment
-    await ChatScreen.uploadFile("./tests/fixtures/logo.jpg");
+    await InputBar.uploadFile("./tests/fixtures/logo.jpg");
 
     // Get the full path of file selected
-    const expectedPath = await ChatScreen.getFilePath(
+    const expectedPath = await InputBar.getFilePath(
       "./tests/fixtures/logo.jpg"
     );
 
     // Validate contents on Compose Attachments are displayed
-    await expect(ChatScreen.composeAttachmentsFileEmbed).toBeDisplayed();
-    await expect(ChatScreen.composeAttachmentsFileIcon).toBeDisplayed();
-    await expect(ChatScreen.composeAttachmentsFileInfo).toBeDisplayed();
+    await expect(ComposeAttachment.composeAttachmentsFileEmbed).toBeDisplayed();
+    await expect(ComposeAttachment.composeAttachmentsFileIcon).toBeDisplayed();
+    await expect(ComposeAttachment.composeAttachmentsFileInfo).toBeDisplayed();
 
     await expect(
-      ChatScreen.composeAttachmentsFileNameText
+      ComposeAttachment.composeAttachmentsFileNameText
     ).toHaveTextContaining(expectedPath);
   });
 
   it("Chats - Delete attachment before sending the message", async () => {
     // Click on upload button and attach a file to compose attachment
-    await ChatScreen.deleteFileOnComposeAttachment();
+    await ComposeAttachment.deleteFileOnComposeAttachment();
 
     // Validate contents on Compose Attachments are displayed
-    await ChatScreen.composeAttachmentsFileEmbed.waitForExist({
+    await ComposeAttachment.composeAttachmentsFileEmbed.waitForExist({
       reverse: true,
     });
   });
 
   it("Chats - Select a file and send message with attachment", async () => {
     // Click on upload button and attach a file to compose attachment
-    await ChatScreen.uploadFile("./tests/fixtures/logo.jpg");
+    await InputBar.uploadFile("./tests/fixtures/logo.jpg");
 
     // Validate contents on Compose Attachments are displayed
-    await expect(ChatScreen.composeAttachmentsFileEmbed).toBeDisplayed();
+    await expect(ComposeAttachment.composeAttachmentsFileEmbed).toBeDisplayed();
 
     // Type a text message and send it
-    await ChatScreen.typeMessageOnInput("message with attachment");
-    await ChatScreen.clickOnSendMessage();
+    await InputBar.typeMessageOnInput("message with attachment");
+    await InputBar.clickOnSendMessage();
   });
 
   it("Chats - Message Sent With Attachment - Text contents", async () => {
     // Validate text from message containing attachment
-    const textFromMessage = await ChatScreen.getLastMessageSentText();
+    const textFromMessage = await Messages.getLastMessageSentText();
     await expect(textFromMessage).toHaveTextContaining(
       "message with attachment"
     );
@@ -256,64 +259,62 @@ xdescribe("Two users at the same time - Chat User A", async () => {
 
   it("Chats - Message Sent With Attachment - File Meta Data", async () => {
     // Validate file metadata is displayed correctly on last chat message sent
-    const fileMeta = await ChatScreen.getLastMessageSentFileMeta();
+    const fileMeta = await Messages.getLastMessageSentFileMeta();
     await expect(fileMeta).toHaveTextContaining("7.75 kB");
   });
 
   it("Chats - Message Sent With Attachment - File Name", async () => {
     // Validate filename is displayed correctly on last chat message sent
-    const fileName = await ChatScreen.getLastMessageSentFileName();
+    const fileName = await Messages.getLastMessageSentFileName();
     await expect(fileName).toHaveTextContaining("logo.jpg");
   });
 
   it("Chats - Message Sent With Attachment - File Icon", async () => {
     // Validate file icon is displayed correctly on last chat message sent
-    const fileIcon = await ChatScreen.getLastMessageSentFileIcon();
+    const fileIcon = await Messages.getLastMessageSentFileIcon();
     await expect(fileIcon).toBeDisplayed();
   });
 
   it("Chats - Message Sent With Attachment - Download Button", async () => {
     // Validate file download button is displayed correctly on last chat message sent
     const fileDownloadButton =
-      await ChatScreen.getLastMessageSentDownloadButton();
+      await Messages.getLastMessageSentDownloadButton();
     await expect(fileDownloadButton).toBeDisplayed();
   });
 
   it("Validate Chat Screen tooltips for Call and Videocall display Coming soon", async () => {
     // Validate Call button tooltip contains "Coming soon"
-    await ChatScreen.hoverOnCallButton();
-    await ChatScreen.topbarCallTooltip.waitForDisplayed();
-    await expect(ChatScreen.topbarCallTooltipText).toHaveTextContaining(
+    await Topbar.hoverOnCallButton();
+    await Topbar.topbarCallTooltip.waitForDisplayed();
+    await expect(Topbar.topbarCallTooltipText).toHaveTextContaining(
       "Coming soon"
     );
 
     // Validate Videocall button tooltip contains "Coming soon"
-    await ChatScreen.hoverOnVideocallButton();
-    await ChatScreen.topbarVideocallTooltip.waitForDisplayed();
-    await expect(ChatScreen.topbarVideocallTooltipText).toHaveTextContaining(
+    await Topbar.hoverOnVideocallButton();
+    await Topbar.topbarVideocallTooltip.waitForDisplayed();
+    await expect(Topbar.topbarVideocallTooltipText).toHaveTextContaining(
       "Coming soon"
     );
   });
 
   it("Validate Chat Screen tooltips are displayed", async () => {
     // Validate Favorites button tooltip
-    await ChatScreen.hoverOnFavoritesButton();
-    await ChatScreen.topbarAddToFavoritesTooltip.waitForDisplayed();
-    await expect(
-      ChatScreen.topbarAddToFavoritesTooltipText
-    ).toHaveTextContaining("Add to Favorites");
+    await Topbar.hoverOnFavoritesButton();
+    await Topbar.topbarAddToFavoritesTooltip.waitForDisplayed();
+    await expect(Topbar.topbarAddToFavoritesTooltipText).toHaveTextContaining(
+      "Add to Favorites"
+    );
 
     // Validate Upload button tooltip
-    await ChatScreen.hoverOnUploadButton();
-    await ChatScreen.uploadTooltip.waitForDisplayed();
-    await expect(ChatScreen.uploadTooltipText).toHaveTextContaining("Upload");
+    await InputBar.hoverOnUploadButton();
+    await InputBar.uploadTooltip.waitForDisplayed();
+    await expect(InputBar.uploadTooltipText).toHaveTextContaining("Upload");
 
     // Validate Send button tooltip
-    await ChatScreen.hoverOnSendButton();
-    await ChatScreen.sendMessageTooltip.waitForDisplayed();
-    await expect(ChatScreen.sendMessageTooltipText).toHaveTextContaining(
-      "Send"
-    );
+    await InputBar.hoverOnSendButton();
+    await InputBar.sendMessageTooltip.waitForDisplayed();
+    await expect(InputBar.sendMessageTooltipText).toHaveTextContaining("Send");
   });
 
   after(async () => {
